# gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package gen

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// N400Error defines model for 400Error.
type N400Error = Error

// N401UnauthorizedError defines model for 401UnauthorizedError.
type N401UnauthorizedError = Error

// N404Error defines model for 404Error.
type N404Error = Error

// N500Error defines model for 500Error.
type N500Error = Error

// CreateUserJSONBody defines parameters for CreateUser.
type CreateUserJSONBody struct {
	FamilyName     string `json:"familyName"`
	FamilyNameKana string `json:"familyNameKana"`
	FirstName      string `json:"firstName"`
	FirstNameKana  string `json:"firstNameKana"`
}

// UpdateUserByIDJSONBody defines parameters for UpdateUserByID.
type UpdateUserByIDJSONBody struct {
	FamilyName     string `json:"familyName"`
	FamilyNameKana string `json:"familyNameKana"`
	FirstName      string `json:"firstName"`
	FirstNameKana  string `json:"firstNameKana"`
}

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody CreateUserJSONBody

// UpdateUserByIDJSONRequestBody defines body for UpdateUserByID for application/json ContentType.
type UpdateUserByIDJSONRequestBody UpdateUserByIDJSONBody